tools/agent-simulator/requirements.txt
python-monascaclient>=1.8.0
